git-svn-id: http://cwinux-mq.googlecode.com/svn/trunk@518 da00a48a-4684-a967-d7f9-89c093a7ad30

+++ mq/zk/zk_get.cpp
@@ -108,9 +108,9 @@ int main(int argc ,char** argv)
 		printf("czxid:%s\n", CwxCommon::toString(stat.czxid, szTmp, 10));
 		printf("mzxid:%s\n", CwxCommon::toString(stat.mzxid, szTmp, 10));
 		timestamp = stat.ctime/1000000;
-		printf("ctime:%d %s", stat.ctime%1000000, ctime(&timestamp));
+		printf("ctime:%d %s", (int)stat.ctime%1000000, ctime(&timestamp));
 		timestamp = stat.mtime/1000000;
-		printf("mtime:%d %s", stat.mtime%1000000, ctime(&timestamp));
+		printf("mtime:%d %s", (int)stat.mtime%1000000, ctime(&timestamp));
 		printf("version:%d\n", stat.version);
 		printf("cversion:%d\n", stat.cversion);
 		printf("aversion:%d\n", stat.aversion);
